requirements.txt
behave==1.2.5
dj-database-url==0.3.0
Django==1.8.7
django-postgrespool==0.3.0
djangorestframework==3.1.1
gunicorn==19.3.0
parse==1.6.6
parse-type==0.3.4
psycopg2==2.6.1
Pygments==2.0.2
requests==2.8.1
six==1.10.0
SQLAlchemy==1.0.8
